+++ packages/plugins/Zoom/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Refactor: Remove redundant prop `isHorizontal` from `Zoom` component. Logic is now placed in `@polar/core`.
 
+- Feature: Add the option to configure a zoom slider.
+
 ## 1.2.0
 
 - Feature: Improved implementation to make plugin SPA-ready.