+++ packages/plugins/Zoom/CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
 ## unpublished
-- Feature: Add new configuration parameter `showZoomSlider` to configure a zoom slider display.
 
+- Feature: Add new configuration parameter `showZoomSlider` to configure a zoom slider display.
 - Refactor: Remove redundant prop `isHorizontal` from `Zoom` component. Logic is now placed in `@polar/core`.
 
 ## 1.2.0