+++ packages/plugins/Zoom/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unpublished
 
+- Feature: Add easing between zoom levels.
+- Feature: Add new configuration parameter `showZoomSlider` to configure a zoom slider display.
 - Refactor: Remove redundant prop `isHorizontal` from `Zoom` component. Logic is now placed in `@polar/core`.
 
 ## 1.2.0